+++ src/storage/storageHeader.ts
@@ -435,7 +435,7 @@ class StorageHeader extends Header {
         if(!Config.get_boolean('storage-header-tooltip'))
             return;
         
-        this.tooltipMenu.open(true);
+        this.tooltipMenu.open(false);
     }
     
     hideTooltip() {
@@ -443,7 +443,7 @@ class StorageHeader extends Header {
             return;
         if(!Config.get_boolean('storage-header-tooltip'))
             return;
-        this.tooltipMenu.close(true);
+        this.tooltipMenu.close(false);
     }
     
     destroy() {
@@ -475,7 +475,7 @@ class StorageHeader extends Header {
         if(this.tooltipMenu) {
             Config.clear(this.tooltipMenu);
             Utils.storageMonitor.unlisten(this.tooltipMenu);
-            this.tooltipMenu.close(true);
+            this.tooltipMenu.close(false);
         }
 
         super.destroy();